Compressed some test files together
